+++ CHANGES.rst
@@ -5,6 +5,18 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Ajaxify cancel bookings.
+  [rnix]
+
+- Fix comment editing in plone 5.
+  [rnix]
+
+- Only display cancel booking action if booking not already cancelled.
+  [rnix]
+
+- Fix error in the order detail if the product no longer exists (#20).
+  [rnix]
+
 - Do not exclude reserved bookings by default from billing. This resolves
   inconsistenty introduced with
   https://github.com/bluedynamics/bda.plone.orders/pull/39 and addressed at